+++ test/unit/SchemaMigrationTest.php
@@ -168,6 +168,7 @@ class SchemaMigrationTest extends TestCase {
 			'wp_sam_change_windows',
 			'wp_sam_tor_exit_nodes',
 			'wp_sam_asn_cache',
+			'wp_sam_geoip_cache',
 		);
 
 		$this->assertSame(